Fix sample scripts
